requirements.txt
alembic==1.9.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
greenlet==2.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
Pillow==9.3.0
psycopg2==2.9.5
SQLAlchemy==1.4.45
Werkzeug==2.2.2
zipp==3.11.0
